[TESTING] Making changes to get django-filemanager working.
These changes were made in an attempt to get the funkybob fork of
django-filemanager working with Python 2.7 and Django 1.8. 3to2 was run
on the entire project resulting in most of the changes seen below. The
only actual manual changes took place in the dispatch function. This is
simply a commit to stash my changes so they are not lost.